JsFighter 2 *BEST BUILD
This build is the latest currently that's the most stable. It still has problems like identifiers, but the images are fixed, The code was massively cleaned up and modified, however the attack button still won't function. We'll see if I get it to work this period or some other time, bur for now unfortinately I couldn't get it to work.